third_party/blink/web_tests/platform/linux/external/wpt/css/css-shapes/shape-outside/values/shape-outside-ellipse-004-expected.txt

This is a testharness.js-based test.
Found 403 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] test unit (inline): cm - ellipse(at 50cm)
  assert_equals: expected "ellipse(at 50cm center)" but got "ellipse(at 50cm 50%)"
[FAIL] test unit (inline): cm - ellipse(at left 50cm)
  assert_equals: expected "ellipse(at left 50cm)" but got "ellipse(at 0% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at 50cm top)
  assert_equals: expected "ellipse(at 50cm top)" but got "ellipse(at 50cm 0%)"
[FAIL] test unit (inline): cm - ellipse(at right 80cm)
  assert_equals: expected "ellipse(at right 80cm)" but got "ellipse(at 100% 80cm)"
[FAIL] test unit (inline): cm - ellipse(at 70cm bottom)
  assert_equals: expected "ellipse(at 70cm bottom)" but got "ellipse(at 70cm 100%)"
[FAIL] test unit (inline): cm - ellipse(at center 60cm)
  assert_equals: expected "ellipse(at center 60cm)" but got "ellipse(at 50% 60cm)"
[FAIL] test unit (inline): cm - ellipse(at 60cm center)
  assert_equals: expected "ellipse(at 60cm center)" but got "ellipse(at 60cm 50%)"
[FAIL] test unit (inline): cm - ellipse(at left 50% top 50cm)
  assert_equals: expected "ellipse(at left 50% top 50cm)" but got "ellipse(at 50% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at left 50cm top 50%)
  assert_equals: expected "ellipse(at left 50cm top 50%)" but got "ellipse(at 50cm 50%)"
[FAIL] test unit (inline): cm - ellipse(at left 50cm top 50cm)
  assert_equals: expected "ellipse(at left 50cm top 50cm)" but got "ellipse(at 50cm 50cm)"
[FAIL] test unit (inline): cm - ellipse(at left 50cm bottom 70%)
  assert_equals: expected "ellipse(at left 50cm bottom 70%)" but got "ellipse(at 50cm 30%)"
[FAIL] test unit (inline): cm - ellipse(at top 50% left 50cm)
  assert_equals: expected "ellipse(at left 50cm top 50%)" but got "ellipse(at 50cm 50%)"
[FAIL] test unit (inline): cm - ellipse(at top 50cm left 50%)
  assert_equals: expected "ellipse(at left 50% top 50cm)" but got "ellipse(at 50% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at top 50cm left 50cm)
  assert_equals: expected "ellipse(at left 50cm top 50cm)" but got "ellipse(at 50cm 50cm)"
[FAIL] test unit (inline): cm - ellipse(at top 50cm right 80%)
  assert_equals: expected "ellipse(at right 80% top 50cm)" but got "ellipse(at 20% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70% left 50cm)
  assert_equals: expected "ellipse(at left 50cm bottom 70%)" but got "ellipse(at 50cm 30%)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70% right 80cm)
  assert_equals: expected "ellipse(at right 80cm bottom 70%)" but got "ellipse(at right 80cm top 30%)"
[FAIL] test unit (inline): cm - ellipse(at bottom 70cm right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70cm)" but got "ellipse(at left 50% bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80% top 50cm)
  assert_equals: expected "ellipse(at right 80% top 50cm)" but got "ellipse(at 20% 50cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80% bottom 70cm)
  assert_equals: expected "ellipse(at right 80% bottom 70cm)" but got "ellipse(at left 20% bottom 70cm)"
[FAIL] test unit (inline): cm - ellipse(at right 80cm bottom 70%)
  assert_equals: expected "ellipse(at right 80cm bottom 70%)" but got "ellipse(at right 80cm top 30%)"
[FAIL] test unit (inline): mm - ellipse(at 50mm)
  assert_equals: expected "ellipse(at 50mm center)" but got "ellipse(at 50mm 50%)"
[FAIL] test unit (inline): mm - ellipse(at left 50mm)
  assert_equals: expected "ellipse(at left 50mm)" but got "ellipse(at 0% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at 50mm top)
  assert_equals: expected "ellipse(at 50mm top)" but got "ellipse(at 50mm 0%)"
[FAIL] test unit (inline): mm - ellipse(at right 80mm)
  assert_equals: expected "ellipse(at right 80mm)" but got "ellipse(at 100% 80mm)"
[FAIL] test unit (inline): mm - ellipse(at 70mm bottom)
  assert_equals: expected "ellipse(at 70mm bottom)" but got "ellipse(at 70mm 100%)"
[FAIL] test unit (inline): mm - ellipse(at center 60mm)
  assert_equals: expected "ellipse(at center 60mm)" but got "ellipse(at 50% 60mm)"
[FAIL] test unit (inline): mm - ellipse(at 60mm center)
  assert_equals: expected "ellipse(at 60mm center)" but got "ellipse(at 60mm 50%)"
[FAIL] test unit (inline): mm - ellipse(at left 50% top 50mm)
  assert_equals: expected "ellipse(at left 50% top 50mm)" but got "ellipse(at 50% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at left 50mm top 50%)
  assert_equals: expected "ellipse(at left 50mm top 50%)" but got "ellipse(at 50mm 50%)"
[FAIL] test unit (inline): mm - ellipse(at left 50mm top 50mm)
  assert_equals: expected "ellipse(at left 50mm top 50mm)" but got "ellipse(at 50mm 50mm)"
[FAIL] test unit (inline): mm - ellipse(at left 50mm bottom 70%)
  assert_equals: expected "ellipse(at left 50mm bottom 70%)" but got "ellipse(at 50mm 30%)"
[FAIL] test unit (inline): mm - ellipse(at top 50% left 50mm)
  assert_equals: expected "ellipse(at left 50mm top 50%)" but got "ellipse(at 50mm 50%)"
[FAIL] test unit (inline): mm - ellipse(at top 50mm left 50%)
  assert_equals: expected "ellipse(at left 50% top 50mm)" but got "ellipse(at 50% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at top 50mm left 50mm)
  assert_equals: expected "ellipse(at left 50mm top 50mm)" but got "ellipse(at 50mm 50mm)"
[FAIL] test unit (inline): mm - ellipse(at top 50mm right 80%)
  assert_equals: expected "ellipse(at right 80% top 50mm)" but got "ellipse(at 20% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70% left 50mm)
  assert_equals: expected "ellipse(at left 50mm bottom 70%)" but got "ellipse(at 50mm 30%)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70% right 80mm)
  assert_equals: expected "ellipse(at right 80mm bottom 70%)" but got "ellipse(at right 80mm top 30%)"
[FAIL] test unit (inline): mm - ellipse(at bottom 70mm right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70mm)" but got "ellipse(at left 50% bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80% top 50mm)
  assert_equals: expected "ellipse(at right 80% top 50mm)" but got "ellipse(at 20% 50mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80% bottom 70mm)
  assert_equals: expected "ellipse(at right 80% bottom 70mm)" but got "ellipse(at left 20% bottom 70mm)"
[FAIL] test unit (inline): mm - ellipse(at right 80mm bottom 70%)
  assert_equals: expected "ellipse(at right 80mm bottom 70%)" but got "ellipse(at right 80mm top 30%)"
[FAIL] test unit (inline): in - ellipse(at 50in)
  assert_equals: expected "ellipse(at 50in center)" but got "ellipse(at 50in 50%)"
[FAIL] test unit (inline): in - ellipse(at left 50in)
  assert_equals: expected "ellipse(at left 50in)" but got "ellipse(at 0% 50in)"
[FAIL] test unit (inline): in - ellipse(at 50in top)
  assert_equals: expected "ellipse(at 50in top)" but got "ellipse(at 50in 0%)"
[FAIL] test unit (inline): in - ellipse(at right 80in)
  assert_equals: expected "ellipse(at right 80in)" but got "ellipse(at 100% 80in)"
[FAIL] test unit (inline): in - ellipse(at 70in bottom)
  assert_equals: expected "ellipse(at 70in bottom)" but got "ellipse(at 70in 100%)"
[FAIL] test unit (inline): in - ellipse(at center 60in)
  assert_equals: expected "ellipse(at center 60in)" but got "ellipse(at 50% 60in)"
[FAIL] test unit (inline): in - ellipse(at 60in center)
  assert_equals: expected "ellipse(at 60in center)" but got "ellipse(at 60in 50%)"
[FAIL] test unit (inline): in - ellipse(at left 50% top 50in)
  assert_equals: expected "ellipse(at left 50% top 50in)" but got "ellipse(at 50% 50in)"
[FAIL] test unit (inline): in - ellipse(at left 50in top 50%)
  assert_equals: expected "ellipse(at left 50in top 50%)" but got "ellipse(at 50in 50%)"
[FAIL] test unit (inline): in - ellipse(at left 50in top 50in)
  assert_equals: expected "ellipse(at left 50in top 50in)" but got "ellipse(at 50in 50in)"
[FAIL] test unit (inline): in - ellipse(at left 50in bottom 70%)
  assert_equals: expected "ellipse(at left 50in bottom 70%)" but got "ellipse(at 50in 30%)"
[FAIL] test unit (inline): in - ellipse(at top 50% left 50in)
  assert_equals: expected "ellipse(at left 50in top 50%)" but got "ellipse(at 50in 50%)"
[FAIL] test unit (inline): in - ellipse(at top 50in left 50%)
  assert_equals: expected "ellipse(at left 50% top 50in)" but got "ellipse(at 50% 50in)"
[FAIL] test unit (inline): in - ellipse(at top 50in left 50in)
  assert_equals: expected "ellipse(at left 50in top 50in)" but got "ellipse(at 50in 50in)"
[FAIL] test unit (inline): in - ellipse(at top 50in right 80%)
  assert_equals: expected "ellipse(at right 80% top 50in)" but got "ellipse(at 20% 50in)"
[FAIL] test unit (inline): in - ellipse(at bottom 70% left 50in)
  assert_equals: expected "ellipse(at left 50in bottom 70%)" but got "ellipse(at 50in 30%)"
[FAIL] test unit (inline): in - ellipse(at bottom 70% right 80in)
  assert_equals: expected "ellipse(at right 80in bottom 70%)" but got "ellipse(at right 80in top 30%)"
[FAIL] test unit (inline): in - ellipse(at bottom 70in right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70in)" but got "ellipse(at left 50% bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at right 80% top 50in)
  assert_equals: expected "ellipse(at right 80% top 50in)" but got "ellipse(at 20% 50in)"
[FAIL] test unit (inline): in - ellipse(at right 80% bottom 70in)
  assert_equals: expected "ellipse(at right 80% bottom 70in)" but got "ellipse(at left 20% bottom 70in)"
[FAIL] test unit (inline): in - ellipse(at right 80in bottom 70%)
  assert_equals: expected "ellipse(at right 80in bottom 70%)" but got "ellipse(at right 80in top 30%)"
[FAIL] test unit (inline): pt - ellipse(at 50pt)
  assert_equals: expected "ellipse(at 50pt center)" but got "ellipse(at 50pt 50%)"
[FAIL] test unit (inline): pt - ellipse(at left 50pt)
  assert_equals: expected "ellipse(at left 50pt)" but got "ellipse(at 0% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at 50pt top)
  assert_equals: expected "ellipse(at 50pt top)" but got "ellipse(at 50pt 0%)"
[FAIL] test unit (inline): pt - ellipse(at right 80pt)
  assert_equals: expected "ellipse(at right 80pt)" but got "ellipse(at 100% 80pt)"
[FAIL] test unit (inline): pt - ellipse(at 70pt bottom)
  assert_equals: expected "ellipse(at 70pt bottom)" but got "ellipse(at 70pt 100%)"
[FAIL] test unit (inline): pt - ellipse(at center 60pt)
  assert_equals: expected "ellipse(at center 60pt)" but got "ellipse(at 50% 60pt)"
[FAIL] test unit (inline): pt - ellipse(at 60pt center)
  assert_equals: expected "ellipse(at 60pt center)" but got "ellipse(at 60pt 50%)"
[FAIL] test unit (inline): pt - ellipse(at left 50% top 50pt)
  assert_equals: expected "ellipse(at left 50% top 50pt)" but got "ellipse(at 50% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at left 50pt top 50%)
  assert_equals: expected "ellipse(at left 50pt top 50%)" but got "ellipse(at 50pt 50%)"
[FAIL] test unit (inline): pt - ellipse(at left 50pt top 50pt)
  assert_equals: expected "ellipse(at left 50pt top 50pt)" but got "ellipse(at 50pt 50pt)"
[FAIL] test unit (inline): pt - ellipse(at left 50pt bottom 70%)
  assert_equals: expected "ellipse(at left 50pt bottom 70%)" but got "ellipse(at 50pt 30%)"
[FAIL] test unit (inline): pt - ellipse(at top 50% left 50pt)
  assert_equals: expected "ellipse(at left 50pt top 50%)" but got "ellipse(at 50pt 50%)"
[FAIL] test unit (inline): pt - ellipse(at top 50pt left 50%)
  assert_equals: expected "ellipse(at left 50% top 50pt)" but got "ellipse(at 50% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at top 50pt left 50pt)
  assert_equals: expected "ellipse(at left 50pt top 50pt)" but got "ellipse(at 50pt 50pt)"
[FAIL] test unit (inline): pt - ellipse(at top 50pt right 80%)
  assert_equals: expected "ellipse(at right 80% top 50pt)" but got "ellipse(at 20% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70% left 50pt)
  assert_equals: expected "ellipse(at left 50pt bottom 70%)" but got "ellipse(at 50pt 30%)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70% right 80pt)
  assert_equals: expected "ellipse(at right 80pt bottom 70%)" but got "ellipse(at right 80pt top 30%)"
[FAIL] test unit (inline): pt - ellipse(at bottom 70pt right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70pt)" but got "ellipse(at left 50% bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80% top 50pt)
  assert_equals: expected "ellipse(at right 80% top 50pt)" but got "ellipse(at 20% 50pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80% bottom 70pt)
  assert_equals: expected "ellipse(at right 80% bottom 70pt)" but got "ellipse(at left 20% bottom 70pt)"
[FAIL] test unit (inline): pt - ellipse(at right 80pt bottom 70%)
  assert_equals: expected "ellipse(at right 80pt bottom 70%)" but got "ellipse(at right 80pt top 30%)"
[FAIL] test unit (inline): pc - ellipse(at 50pc)
  assert_equals: expected "ellipse(at 50pc center)" but got "ellipse(at 50pc 50%)"
[FAIL] test unit (inline): pc - ellipse(at left 50pc)
  assert_equals: expected "ellipse(at left 50pc)" but got "ellipse(at 0% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at 50pc top)
  assert_equals: expected "ellipse(at 50pc top)" but got "ellipse(at 50pc 0%)"
[FAIL] test unit (inline): pc - ellipse(at right 80pc)
  assert_equals: expected "ellipse(at right 80pc)" but got "ellipse(at 100% 80pc)"
[FAIL] test unit (inline): pc - ellipse(at 70pc bottom)
  assert_equals: expected "ellipse(at 70pc bottom)" but got "ellipse(at 70pc 100%)"
[FAIL] test unit (inline): pc - ellipse(at center 60pc)
  assert_equals: expected "ellipse(at center 60pc)" but got "ellipse(at 50% 60pc)"
[FAIL] test unit (inline): pc - ellipse(at 60pc center)
  assert_equals: expected "ellipse(at 60pc center)" but got "ellipse(at 60pc 50%)"
[FAIL] test unit (inline): pc - ellipse(at left 50% top 50pc)
  assert_equals: expected "ellipse(at left 50% top 50pc)" but got "ellipse(at 50% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at left 50pc top 50%)
  assert_equals: expected "ellipse(at left 50pc top 50%)" but got "ellipse(at 50pc 50%)"
[FAIL] test unit (inline): pc - ellipse(at left 50pc top 50pc)
  assert_equals: expected "ellipse(at left 50pc top 50pc)" but got "ellipse(at 50pc 50pc)"
[FAIL] test unit (inline): pc - ellipse(at left 50pc bottom 70%)
  assert_equals: expected "ellipse(at left 50pc bottom 70%)" but got "ellipse(at 50pc 30%)"
[FAIL] test unit (inline): pc - ellipse(at top 50% left 50pc)
  assert_equals: expected "ellipse(at left 50pc top 50%)" but got "ellipse(at 50pc 50%)"
[FAIL] test unit (inline): pc - ellipse(at top 50pc left 50%)
  assert_equals: expected "ellipse(at left 50% top 50pc)" but got "ellipse(at 50% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at top 50pc left 50pc)
  assert_equals: expected "ellipse(at left 50pc top 50pc)" but got "ellipse(at 50pc 50pc)"
[FAIL] test unit (inline): pc - ellipse(at top 50pc right 80%)
  assert_equals: expected "ellipse(at right 80% top 50pc)" but got "ellipse(at 20% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70% left 50pc)
  assert_equals: expected "ellipse(at left 50pc bottom 70%)" but got "ellipse(at 50pc 30%)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70% right 80pc)
  assert_equals: expected "ellipse(at right 80pc bottom 70%)" but got "ellipse(at right 80pc top 30%)"
[FAIL] test unit (inline): pc - ellipse(at bottom 70pc right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70pc)" but got "ellipse(at left 50% bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80% top 50pc)
  assert_equals: expected "ellipse(at right 80% top 50pc)" but got "ellipse(at 20% 50pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80% bottom 70pc)
  assert_equals: expected "ellipse(at right 80% bottom 70pc)" but got "ellipse(at left 20% bottom 70pc)"
[FAIL] test unit (inline): pc - ellipse(at right 80pc bottom 70%)
  assert_equals: expected "ellipse(at right 80pc bottom 70%)" but got "ellipse(at right 80pc top 30%)"
[FAIL] test unit (inline): em - ellipse(at 50em)
  assert_equals: expected "ellipse(at 50em center)" but got "ellipse(at 50em 50%)"
[FAIL] test unit (inline): em - ellipse(at left 50em)
  assert_equals: expected "ellipse(at left 50em)" but got "ellipse(at 0% 50em)"
[FAIL] test unit (inline): em - ellipse(at 50em top)
  assert_equals: expected "ellipse(at 50em top)" but got "ellipse(at 50em 0%)"
[FAIL] test unit (inline): em - ellipse(at right 80em)
  assert_equals: expected "ellipse(at right 80em)" but got "ellipse(at 100% 80em)"
[FAIL] test unit (inline): em - ellipse(at 70em bottom)
  assert_equals: expected "ellipse(at 70em bottom)" but got "ellipse(at 70em 100%)"
[FAIL] test unit (inline): em - ellipse(at center 60em)
  assert_equals: expected "ellipse(at center 60em)" but got "ellipse(at 50% 60em)"
[FAIL] test unit (inline): em - ellipse(at 60em center)
  assert_equals: expected "ellipse(at 60em center)" but got "ellipse(at 60em 50%)"
[FAIL] test unit (inline): em - ellipse(at left 50% top 50em)
  assert_equals: expected "ellipse(at left 50% top 50em)" but got "ellipse(at 50% 50em)"
[FAIL] test unit (inline): em - ellipse(at left 50em top 50%)
  assert_equals: expected "ellipse(at left 50em top 50%)" but got "ellipse(at 50em 50%)"
[FAIL] test unit (inline): em - ellipse(at left 50em top 50em)
  assert_equals: expected "ellipse(at left 50em top 50em)" but got "ellipse(at 50em 50em)"
[FAIL] test unit (inline): em - ellipse(at left 50em bottom 70%)
  assert_equals: expected "ellipse(at left 50em bottom 70%)" but got "ellipse(at 50em 30%)"
[FAIL] test unit (inline): em - ellipse(at top 50% left 50em)
  assert_equals: expected "ellipse(at left 50em top 50%)" but got "ellipse(at 50em 50%)"
[FAIL] test unit (inline): em - ellipse(at top 50em left 50%)
  assert_equals: expected "ellipse(at left 50% top 50em)" but got "ellipse(at 50% 50em)"
[FAIL] test unit (inline): em - ellipse(at top 50em left 50em)
  assert_equals: expected "ellipse(at left 50em top 50em)" but got "ellipse(at 50em 50em)"
[FAIL] test unit (inline): em - ellipse(at top 50em right 80%)
  assert_equals: expected "ellipse(at right 80% top 50em)" but got "ellipse(at 20% 50em)"
[FAIL] test unit (inline): em - ellipse(at bottom 70% left 50em)
  assert_equals: expected "ellipse(at left 50em bottom 70%)" but got "ellipse(at 50em 30%)"
[FAIL] test unit (inline): em - ellipse(at bottom 70% right 80em)
  assert_equals: expected "ellipse(at right 80em bottom 70%)" but got "ellipse(at right 80em top 30%)"
[FAIL] test unit (inline): em - ellipse(at bottom 70em right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70em)" but got "ellipse(at left 50% bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at right 80% top 50em)
  assert_equals: expected "ellipse(at right 80% top 50em)" but got "ellipse(at 20% 50em)"
[FAIL] test unit (inline): em - ellipse(at right 80% bottom 70em)
  assert_equals: expected "ellipse(at right 80% bottom 70em)" but got "ellipse(at left 20% bottom 70em)"
[FAIL] test unit (inline): em - ellipse(at right 80em bottom 70%)
  assert_equals: expected "ellipse(at right 80em bottom 70%)" but got "ellipse(at right 80em top 30%)"
[FAIL] test unit (inline): ex - ellipse(at 50ex)
  assert_equals: expected "ellipse(at 50ex center)" but got "ellipse(at 50ex 50%)"
[FAIL] test unit (inline): ex - ellipse(at left 50ex)
  assert_equals: expected "ellipse(at left 50ex)" but got "ellipse(at 0% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at 50ex top)
  assert_equals: expected "ellipse(at 50ex top)" but got "ellipse(at 50ex 0%)"
[FAIL] test unit (inline): ex - ellipse(at right 80ex)
  assert_equals: expected "ellipse(at right 80ex)" but got "ellipse(at 100% 80ex)"
[FAIL] test unit (inline): ex - ellipse(at 70ex bottom)
  assert_equals: expected "ellipse(at 70ex bottom)" but got "ellipse(at 70ex 100%)"
[FAIL] test unit (inline): ex - ellipse(at center 60ex)
  assert_equals: expected "ellipse(at center 60ex)" but got "ellipse(at 50% 60ex)"
[FAIL] test unit (inline): ex - ellipse(at 60ex center)
  assert_equals: expected "ellipse(at 60ex center)" but got "ellipse(at 60ex 50%)"
[FAIL] test unit (inline): ex - ellipse(at left 50% top 50ex)
  assert_equals: expected "ellipse(at left 50% top 50ex)" but got "ellipse(at 50% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at left 50ex top 50%)
  assert_equals: expected "ellipse(at left 50ex top 50%)" but got "ellipse(at 50ex 50%)"
[FAIL] test unit (inline): ex - ellipse(at left 50ex top 50ex)
  assert_equals: expected "ellipse(at left 50ex top 50ex)" but got "ellipse(at 50ex 50ex)"
[FAIL] test unit (inline): ex - ellipse(at left 50ex bottom 70%)
  assert_equals: expected "ellipse(at left 50ex bottom 70%)" but got "ellipse(at 50ex 30%)"
[FAIL] test unit (inline): ex - ellipse(at top 50% left 50ex)
  assert_equals: expected "ellipse(at left 50ex top 50%)" but got "ellipse(at 50ex 50%)"
[FAIL] test unit (inline): ex - ellipse(at top 50ex left 50%)
  assert_equals: expected "ellipse(at left 50% top 50ex)" but got "ellipse(at 50% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at top 50ex left 50ex)
  assert_equals: expected "ellipse(at left 50ex top 50ex)" but got "ellipse(at 50ex 50ex)"
[FAIL] test unit (inline): ex - ellipse(at top 50ex right 80%)
  assert_equals: expected "ellipse(at right 80% top 50ex)" but got "ellipse(at 20% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70% left 50ex)
  assert_equals: expected "ellipse(at left 50ex bottom 70%)" but got "ellipse(at 50ex 30%)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70% right 80ex)
  assert_equals: expected "ellipse(at right 80ex bottom 70%)" but got "ellipse(at right 80ex top 30%)"
[FAIL] test unit (inline): ex - ellipse(at bottom 70ex right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70ex)" but got "ellipse(at left 50% bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80% top 50ex)
  assert_equals: expected "ellipse(at right 80% top 50ex)" but got "ellipse(at 20% 50ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80% bottom 70ex)
  assert_equals: expected "ellipse(at right 80% bottom 70ex)" but got "ellipse(at left 20% bottom 70ex)"
[FAIL] test unit (inline): ex - ellipse(at right 80ex bottom 70%)
  assert_equals: expected "ellipse(at right 80ex bottom 70%)" but got "ellipse(at right 80ex top 30%)"
[FAIL] test unit (inline): ch - ellipse(at 50ch)
  assert_equals: expected "ellipse(at 50ch center)" but got "ellipse(at 50ch 50%)"
[FAIL] test unit (inline): ch - ellipse(at left 50ch)
  assert_equals: expected "ellipse(at left 50ch)" but got "ellipse(at 0% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at 50ch top)
  assert_equals: expected "ellipse(at 50ch top)" but got "ellipse(at 50ch 0%)"
[FAIL] test unit (inline): ch - ellipse(at right 80ch)
  assert_equals: expected "ellipse(at right 80ch)" but got "ellipse(at 100% 80ch)"
[FAIL] test unit (inline): ch - ellipse(at 70ch bottom)
  assert_equals: expected "ellipse(at 70ch bottom)" but got "ellipse(at 70ch 100%)"
[FAIL] test unit (inline): ch - ellipse(at center 60ch)
  assert_equals: expected "ellipse(at center 60ch)" but got "ellipse(at 50% 60ch)"
[FAIL] test unit (inline): ch - ellipse(at 60ch center)
  assert_equals: expected "ellipse(at 60ch center)" but got "ellipse(at 60ch 50%)"
[FAIL] test unit (inline): ch - ellipse(at left 50% top 50ch)
  assert_equals: expected "ellipse(at left 50% top 50ch)" but got "ellipse(at 50% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at left 50ch top 50%)
  assert_equals: expected "ellipse(at left 50ch top 50%)" but got "ellipse(at 50ch 50%)"
[FAIL] test unit (inline): ch - ellipse(at left 50ch top 50ch)
  assert_equals: expected "ellipse(at left 50ch top 50ch)" but got "ellipse(at 50ch 50ch)"
[FAIL] test unit (inline): ch - ellipse(at left 50ch bottom 70%)
  assert_equals: expected "ellipse(at left 50ch bottom 70%)" but got "ellipse(at 50ch 30%)"
[FAIL] test unit (inline): ch - ellipse(at top 50% left 50ch)
  assert_equals: expected "ellipse(at left 50ch top 50%)" but got "ellipse(at 50ch 50%)"
[FAIL] test unit (inline): ch - ellipse(at top 50ch left 50%)
  assert_equals: expected "ellipse(at left 50% top 50ch)" but got "ellipse(at 50% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at top 50ch left 50ch)
  assert_equals: expected "ellipse(at left 50ch top 50ch)" but got "ellipse(at 50ch 50ch)"
[FAIL] test unit (inline): ch - ellipse(at top 50ch right 80%)
  assert_equals: expected "ellipse(at right 80% top 50ch)" but got "ellipse(at 20% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70% left 50ch)
  assert_equals: expected "ellipse(at left 50ch bottom 70%)" but got "ellipse(at 50ch 30%)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70% right 80ch)
  assert_equals: expected "ellipse(at right 80ch bottom 70%)" but got "ellipse(at right 80ch top 30%)"
[FAIL] test unit (inline): ch - ellipse(at bottom 70ch right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70ch)" but got "ellipse(at left 50% bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80% top 50ch)
  assert_equals: expected "ellipse(at right 80% top 50ch)" but got "ellipse(at 20% 50ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80% bottom 70ch)
  assert_equals: expected "ellipse(at right 80% bottom 70ch)" but got "ellipse(at left 20% bottom 70ch)"
[FAIL] test unit (inline): ch - ellipse(at right 80ch bottom 70%)
  assert_equals: expected "ellipse(at right 80ch bottom 70%)" but got "ellipse(at right 80ch top 30%)"
[FAIL] test unit (inline): rem - ellipse(at 50rem)
  assert_equals: expected "ellipse(at 50rem center)" but got "ellipse(at 50rem 50%)"
[FAIL] test unit (inline): rem - ellipse(at left 50rem)
  assert_equals: expected "ellipse(at left 50rem)" but got "ellipse(at 0% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at 50rem top)
  assert_equals: expected "ellipse(at 50rem top)" but got "ellipse(at 50rem 0%)"
[FAIL] test unit (inline): rem - ellipse(at right 80rem)
  assert_equals: expected "ellipse(at right 80rem)" but got "ellipse(at 100% 80rem)"
[FAIL] test unit (inline): rem - ellipse(at 70rem bottom)
  assert_equals: expected "ellipse(at 70rem bottom)" but got "ellipse(at 70rem 100%)"
[FAIL] test unit (inline): rem - ellipse(at center 60rem)
  assert_equals: expected "ellipse(at center 60rem)" but got "ellipse(at 50% 60rem)"
[FAIL] test unit (inline): rem - ellipse(at 60rem center)
  assert_equals: expected "ellipse(at 60rem center)" but got "ellipse(at 60rem 50%)"
[FAIL] test unit (inline): rem - ellipse(at left 50% top 50rem)
  assert_equals: expected "ellipse(at left 50% top 50rem)" but got "ellipse(at 50% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at left 50rem top 50%)
  assert_equals: expected "ellipse(at left 50rem top 50%)" but got "ellipse(at 50rem 50%)"
[FAIL] test unit (inline): rem - ellipse(at left 50rem top 50rem)
  assert_equals: expected "ellipse(at left 50rem top 50rem)" but got "ellipse(at 50rem 50rem)"
[FAIL] test unit (inline): rem - ellipse(at left 50rem bottom 70%)
  assert_equals: expected "ellipse(at left 50rem bottom 70%)" but got "ellipse(at 50rem 30%)"
[FAIL] test unit (inline): rem - ellipse(at top 50% left 50rem)
  assert_equals: expected "ellipse(at left 50rem top 50%)" but got "ellipse(at 50rem 50%)"
[FAIL] test unit (inline): rem - ellipse(at top 50rem left 50%)
  assert_equals: expected "ellipse(at left 50% top 50rem)" but got "ellipse(at 50% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at top 50rem left 50rem)
  assert_equals: expected "ellipse(at left 50rem top 50rem)" but got "ellipse(at 50rem 50rem)"
[FAIL] test unit (inline): rem - ellipse(at top 50rem right 80%)
  assert_equals: expected "ellipse(at right 80% top 50rem)" but got "ellipse(at 20% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70% left 50rem)
  assert_equals: expected "ellipse(at left 50rem bottom 70%)" but got "ellipse(at 50rem 30%)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70% right 80rem)
  assert_equals: expected "ellipse(at right 80rem bottom 70%)" but got "ellipse(at right 80rem top 30%)"
[FAIL] test unit (inline): rem - ellipse(at bottom 70rem right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70rem)" but got "ellipse(at left 50% bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80% top 50rem)
  assert_equals: expected "ellipse(at right 80% top 50rem)" but got "ellipse(at 20% 50rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80% bottom 70rem)
  assert_equals: expected "ellipse(at right 80% bottom 70rem)" but got "ellipse(at left 20% bottom 70rem)"
[FAIL] test unit (inline): rem - ellipse(at right 80rem bottom 70%)
  assert_equals: expected "ellipse(at right 80rem bottom 70%)" but got "ellipse(at right 80rem top 30%)"
[FAIL] test unit (inline): vw - ellipse(at 50vw)
  assert_equals: expected "ellipse(at 50vw center)" but got "ellipse(at 50vw 50%)"
[FAIL] test unit (inline): vw - ellipse(at left 50vw)
  assert_equals: expected "ellipse(at left 50vw)" but got "ellipse(at 0% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at 50vw top)
  assert_equals: expected "ellipse(at 50vw top)" but got "ellipse(at 50vw 0%)"
[FAIL] test unit (inline): vw - ellipse(at right 80vw)
  assert_equals: expected "ellipse(at right 80vw)" but got "ellipse(at 100% 80vw)"
[FAIL] test unit (inline): vw - ellipse(at 70vw bottom)
  assert_equals: expected "ellipse(at 70vw bottom)" but got "ellipse(at 70vw 100%)"
[FAIL] test unit (inline): vw - ellipse(at center 60vw)
  assert_equals: expected "ellipse(at center 60vw)" but got "ellipse(at 50% 60vw)"
[FAIL] test unit (inline): vw - ellipse(at 60vw center)
  assert_equals: expected "ellipse(at 60vw center)" but got "ellipse(at 60vw 50%)"
[FAIL] test unit (inline): vw - ellipse(at left 50% top 50vw)
  assert_equals: expected "ellipse(at left 50% top 50vw)" but got "ellipse(at 50% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at left 50vw top 50%)
  assert_equals: expected "ellipse(at left 50vw top 50%)" but got "ellipse(at 50vw 50%)"
[FAIL] test unit (inline): vw - ellipse(at left 50vw top 50vw)
  assert_equals: expected "ellipse(at left 50vw top 50vw)" but got "ellipse(at 50vw 50vw)"
[FAIL] test unit (inline): vw - ellipse(at left 50vw bottom 70%)
  assert_equals: expected "ellipse(at left 50vw bottom 70%)" but got "ellipse(at 50vw 30%)"
[FAIL] test unit (inline): vw - ellipse(at top 50% left 50vw)
  assert_equals: expected "ellipse(at left 50vw top 50%)" but got "ellipse(at 50vw 50%)"
[FAIL] test unit (inline): vw - ellipse(at top 50vw left 50%)
  assert_equals: expected "ellipse(at left 50% top 50vw)" but got "ellipse(at 50% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at top 50vw left 50vw)
  assert_equals: expected "ellipse(at left 50vw top 50vw)" but got "ellipse(at 50vw 50vw)"
[FAIL] test unit (inline): vw - ellipse(at top 50vw right 80%)
  assert_equals: expected "ellipse(at right 80% top 50vw)" but got "ellipse(at 20% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70% left 50vw)
  assert_equals: expected "ellipse(at left 50vw bottom 70%)" but got "ellipse(at 50vw 30%)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70% right 80vw)
  assert_equals: expected "ellipse(at right 80vw bottom 70%)" but got "ellipse(at right 80vw top 30%)"
[FAIL] test unit (inline): vw - ellipse(at bottom 70vw right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70vw)" but got "ellipse(at left 50% bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80% top 50vw)
  assert_equals: expected "ellipse(at right 80% top 50vw)" but got "ellipse(at 20% 50vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80% bottom 70vw)
  assert_equals: expected "ellipse(at right 80% bottom 70vw)" but got "ellipse(at left 20% bottom 70vw)"
[FAIL] test unit (inline): vw - ellipse(at right 80vw bottom 70%)
  assert_equals: expected "ellipse(at right 80vw bottom 70%)" but got "ellipse(at right 80vw top 30%)"
[FAIL] test unit (inline): vh - ellipse(at 50vh)
  assert_equals: expected "ellipse(at 50vh center)" but got "ellipse(at 50vh 50%)"
[FAIL] test unit (inline): vh - ellipse(at left 50vh)
  assert_equals: expected "ellipse(at left 50vh)" but got "ellipse(at 0% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at 50vh top)
  assert_equals: expected "ellipse(at 50vh top)" but got "ellipse(at 50vh 0%)"
[FAIL] test unit (inline): vh - ellipse(at right 80vh)
  assert_equals: expected "ellipse(at right 80vh)" but got "ellipse(at 100% 80vh)"
[FAIL] test unit (inline): vh - ellipse(at 70vh bottom)
  assert_equals: expected "ellipse(at 70vh bottom)" but got "ellipse(at 70vh 100%)"
[FAIL] test unit (inline): vh - ellipse(at center 60vh)
  assert_equals: expected "ellipse(at center 60vh)" but got "ellipse(at 50% 60vh)"
[FAIL] test unit (inline): vh - ellipse(at 60vh center)
  assert_equals: expected "ellipse(at 60vh center)" but got "ellipse(at 60vh 50%)"
[FAIL] test unit (inline): vh - ellipse(at left 50% top 50vh)
  assert_equals: expected "ellipse(at left 50% top 50vh)" but got "ellipse(at 50% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at left 50vh top 50%)
  assert_equals: expected "ellipse(at left 50vh top 50%)" but got "ellipse(at 50vh 50%)"
[FAIL] test unit (inline): vh - ellipse(at left 50vh top 50vh)
  assert_equals: expected "ellipse(at left 50vh top 50vh)" but got "ellipse(at 50vh 50vh)"
[FAIL] test unit (inline): vh - ellipse(at left 50vh bottom 70%)
  assert_equals: expected "ellipse(at left 50vh bottom 70%)" but got "ellipse(at 50vh 30%)"
[FAIL] test unit (inline): vh - ellipse(at top 50% left 50vh)
  assert_equals: expected "ellipse(at left 50vh top 50%)" but got "ellipse(at 50vh 50%)"
[FAIL] test unit (inline): vh - ellipse(at top 50vh left 50%)
  assert_equals: expected "ellipse(at left 50% top 50vh)" but got "ellipse(at 50% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at top 50vh left 50vh)
  assert_equals: expected "ellipse(at left 50vh top 50vh)" but got "ellipse(at 50vh 50vh)"
[FAIL] test unit (inline): vh - ellipse(at top 50vh right 80%)
  assert_equals: expected "ellipse(at right 80% top 50vh)" but got "ellipse(at 20% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70% left 50vh)
  assert_equals: expected "ellipse(at left 50vh bottom 70%)" but got "ellipse(at 50vh 30%)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70% right 80vh)
  assert_equals: expected "ellipse(at right 80vh bottom 70%)" but got "ellipse(at right 80vh top 30%)"
[FAIL] test unit (inline): vh - ellipse(at bottom 70vh right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70vh)" but got "ellipse(at left 50% bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80% top 50vh)
  assert_equals: expected "ellipse(at right 80% top 50vh)" but got "ellipse(at 20% 50vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80% bottom 70vh)
  assert_equals: expected "ellipse(at right 80% bottom 70vh)" but got "ellipse(at left 20% bottom 70vh)"
[FAIL] test unit (inline): vh - ellipse(at right 80vh bottom 70%)
  assert_equals: expected "ellipse(at right 80vh bottom 70%)" but got "ellipse(at right 80vh top 30%)"
[FAIL] test unit (inline): vmin - ellipse(at 50vmin)
  assert_equals: expected "ellipse(at 50vmin center)" but got "ellipse(at 50vmin 50%)"
[FAIL] test unit (inline): vmin - ellipse(at left 50vmin)
  assert_equals: expected "ellipse(at left 50vmin)" but got "ellipse(at 0% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at 50vmin top)
  assert_equals: expected "ellipse(at 50vmin top)" but got "ellipse(at 50vmin 0%)"
[FAIL] test unit (inline): vmin - ellipse(at right 80vmin)
  assert_equals: expected "ellipse(at right 80vmin)" but got "ellipse(at 100% 80vmin)"
[FAIL] test unit (inline): vmin - ellipse(at 70vmin bottom)
  assert_equals: expected "ellipse(at 70vmin bottom)" but got "ellipse(at 70vmin 100%)"
[FAIL] test unit (inline): vmin - ellipse(at center 60vmin)
  assert_equals: expected "ellipse(at center 60vmin)" but got "ellipse(at 50% 60vmin)"
[FAIL] test unit (inline): vmin - ellipse(at 60vmin center)
  assert_equals: expected "ellipse(at 60vmin center)" but got "ellipse(at 60vmin 50%)"
[FAIL] test unit (inline): vmin - ellipse(at left 50% top 50vmin)
  assert_equals: expected "ellipse(at left 50% top 50vmin)" but got "ellipse(at 50% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at left 50vmin top 50%)
  assert_equals: expected "ellipse(at left 50vmin top 50%)" but got "ellipse(at 50vmin 50%)"
[FAIL] test unit (inline): vmin - ellipse(at left 50vmin top 50vmin)
  assert_equals: expected "ellipse(at left 50vmin top 50vmin)" but got "ellipse(at 50vmin 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at left 50vmin bottom 70%)
  assert_equals: expected "ellipse(at left 50vmin bottom 70%)" but got "ellipse(at 50vmin 30%)"
[FAIL] test unit (inline): vmin - ellipse(at top 50% left 50vmin)
  assert_equals: expected "ellipse(at left 50vmin top 50%)" but got "ellipse(at 50vmin 50%)"
[FAIL] test unit (inline): vmin - ellipse(at top 50vmin left 50%)
  assert_equals: expected "ellipse(at left 50% top 50vmin)" but got "ellipse(at 50% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at top 50vmin left 50vmin)
  assert_equals: expected "ellipse(at left 50vmin top 50vmin)" but got "ellipse(at 50vmin 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at top 50vmin right 80%)
  assert_equals: expected "ellipse(at right 80% top 50vmin)" but got "ellipse(at 20% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70% left 50vmin)
  assert_equals: expected "ellipse(at left 50vmin bottom 70%)" but got "ellipse(at 50vmin 30%)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70% right 80vmin)
  assert_equals: expected "ellipse(at right 80vmin bottom 70%)" but got "ellipse(at right 80vmin top 30%)"
[FAIL] test unit (inline): vmin - ellipse(at bottom 70vmin right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70vmin)" but got "ellipse(at left 50% bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80% top 50vmin)
  assert_equals: expected "ellipse(at right 80% top 50vmin)" but got "ellipse(at 20% 50vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80% bottom 70vmin)
  assert_equals: expected "ellipse(at right 80% bottom 70vmin)" but got "ellipse(at left 20% bottom 70vmin)"
[FAIL] test unit (inline): vmin - ellipse(at right 80vmin bottom 70%)
  assert_equals: expected "ellipse(at right 80vmin bottom 70%)" but got "ellipse(at right 80vmin top 30%)"
[FAIL] test unit (inline): vmax - ellipse(at 50vmax)
  assert_equals: expected "ellipse(at 50vmax center)" but got "ellipse(at 50vmax 50%)"
[FAIL] test unit (inline): vmax - ellipse(at left 50vmax)
  assert_equals: expected "ellipse(at left 50vmax)" but got "ellipse(at 0% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at 50vmax top)
  assert_equals: expected "ellipse(at 50vmax top)" but got "ellipse(at 50vmax 0%)"
[FAIL] test unit (inline): vmax - ellipse(at right 80vmax)
  assert_equals: expected "ellipse(at right 80vmax)" but got "ellipse(at 100% 80vmax)"
[FAIL] test unit (inline): vmax - ellipse(at 70vmax bottom)
  assert_equals: expected "ellipse(at 70vmax bottom)" but got "ellipse(at 70vmax 100%)"
[FAIL] test unit (inline): vmax - ellipse(at center 60vmax)
  assert_equals: expected "ellipse(at center 60vmax)" but got "ellipse(at 50% 60vmax)"
[FAIL] test unit (inline): vmax - ellipse(at 60vmax center)
  assert_equals: expected "ellipse(at 60vmax center)" but got "ellipse(at 60vmax 50%)"
[FAIL] test unit (inline): vmax - ellipse(at left 50% top 50vmax)
  assert_equals: expected "ellipse(at left 50% top 50vmax)" but got "ellipse(at 50% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at left 50vmax top 50%)
  assert_equals: expected "ellipse(at left 50vmax top 50%)" but got "ellipse(at 50vmax 50%)"
[FAIL] test unit (inline): vmax - ellipse(at left 50vmax top 50vmax)
  assert_equals: expected "ellipse(at left 50vmax top 50vmax)" but got "ellipse(at 50vmax 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at left 50vmax bottom 70%)
  assert_equals: expected "ellipse(at left 50vmax bottom 70%)" but got "ellipse(at 50vmax 30%)"
[FAIL] test unit (inline): vmax - ellipse(at top 50% left 50vmax)
  assert_equals: expected "ellipse(at left 50vmax top 50%)" but got "ellipse(at 50vmax 50%)"
[FAIL] test unit (inline): vmax - ellipse(at top 50vmax left 50%)
  assert_equals: expected "ellipse(at left 50% top 50vmax)" but got "ellipse(at 50% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at top 50vmax left 50vmax)
  assert_equals: expected "ellipse(at left 50vmax top 50vmax)" but got "ellipse(at 50vmax 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at top 50vmax right 80%)
  assert_equals: expected "ellipse(at right 80% top 50vmax)" but got "ellipse(at 20% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70% left 50vmax)
  assert_equals: expected "ellipse(at left 50vmax bottom 70%)" but got "ellipse(at 50vmax 30%)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70% right 80vmax)
  assert_equals: expected "ellipse(at right 80vmax bottom 70%)" but got "ellipse(at right 80vmax top 30%)"
[FAIL] test unit (inline): vmax - ellipse(at bottom 70vmax right 50%)
  assert_equals: expected "ellipse(at right 50% bottom 70vmax)" but got "ellipse(at left 50% bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80% top 50vmax)
  assert_equals: expected "ellipse(at right 80% top 50vmax)" but got "ellipse(at 20% 50vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80% bottom 70vmax)
  assert_equals: expected "ellipse(at right 80% bottom 70vmax)" but got "ellipse(at left 20% bottom 70vmax)"
[FAIL] test unit (inline): vmax - ellipse(at right 80vmax bottom 70%)
  assert_equals: expected "ellipse(at right 80vmax bottom 70%)" but got "ellipse(at right 80vmax top 30%)"
[FAIL] test unit (computed): cm - ellipse(at left 50% bottom 70cm)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at top 50% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 3023.62px top 50%)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70% right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 30%)" but got "ellipse(at right 3023.62px top 30%)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 2645.67px))" but got "ellipse(at left 50% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at bottom 70cm right 80cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 3023.62px bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at right 80% bottom 70cm)
  assert_equals: expected "ellipse(at 20% calc(100% - 2645.67px))" but got "ellipse(at left 20% bottom 2645.67px)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 50%)" but got "ellipse(at right 3023.62px top 50%)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) 30%)" but got "ellipse(at right 3023.62px top 30%)"
[FAIL] test unit (computed): cm - ellipse(at right 80cm bottom 70cm)
  assert_equals: expected "ellipse(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "ellipse(at right 3023.62px bottom 2645.67px)"
[FAIL] test unit (computed): mm - ellipse(at left 50% bottom 70mm)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at top 50% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 302.362px top 50%)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70% right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 30%)" but got "ellipse(at right 302.362px top 30%)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 264.567px))" but got "ellipse(at left 50% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at bottom 70mm right 80mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 302.362px bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at right 80% bottom 70mm)
  assert_equals: expected "ellipse(at 20% calc(100% - 264.567px))" but got "ellipse(at left 20% bottom 264.567px)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm top 50%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 50%)" but got "ellipse(at right 302.362px top 50%)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) 30%)" but got "ellipse(at right 302.362px top 30%)"
[FAIL] test unit (computed): mm - ellipse(at right 80mm bottom 70mm)
  assert_equals: expected "ellipse(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "ellipse(at right 302.362px bottom 264.567px)"
[FAIL] test unit (computed): in - ellipse(at left 50% bottom 70in)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at top 50% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 7680px top 50%)"
[FAIL] test unit (computed): in - ellipse(at bottom 70% right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 30%)" but got "ellipse(at right 7680px top 30%)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 6720px))" but got "ellipse(at left 50% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at bottom 70in right 80in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 7680px bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at right 80% bottom 70in)
  assert_equals: expected "ellipse(at 20% calc(100% - 6720px))" but got "ellipse(at left 20% bottom 6720px)"
[FAIL] test unit (computed): in - ellipse(at right 80in top 50%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 50%)" but got "ellipse(at right 7680px top 50%)"
[FAIL] test unit (computed): in - ellipse(at right 80in bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 7680px) 30%)" but got "ellipse(at right 7680px top 30%)"
[FAIL] test unit (computed): in - ellipse(at right 80in bottom 70in)
  assert_equals: expected "ellipse(at calc(100% - 7680px) calc(100% - 6720px))" but got "ellipse(at right 7680px bottom 6720px)"
[FAIL] test unit (computed): pt - ellipse(at left 50% bottom 70pt)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at top 50% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 106.667px top 50%)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70% right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 30%)" but got "ellipse(at right 106.667px top 30%)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 93.3333px))" but got "ellipse(at left 50% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at bottom 70pt right 80pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 106.667px bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at right 80% bottom 70pt)
  assert_equals: expected "ellipse(at 20% calc(100% - 93.3333px))" but got "ellipse(at left 20% bottom 93.3333px)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt top 50%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 50%)" but got "ellipse(at right 106.667px top 50%)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) 30%)" but got "ellipse(at right 106.667px top 30%)"
[FAIL] test unit (computed): pt - ellipse(at right 80pt bottom 70pt)
  assert_equals: expected "ellipse(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "ellipse(at right 106.667px bottom 93.3333px)"
[FAIL] test unit (computed): pc - ellipse(at left 50% bottom 70pc)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at top 50% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70% right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at bottom 70pc right 80pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at right 80% bottom 70pc)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): pc - ellipse(at right 80pc bottom 70pc)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at left 50% bottom 70em)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at top 50% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): em - ellipse(at bottom 70% right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at bottom 70em right 80em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at right 80% bottom 70em)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): em - ellipse(at right 80em top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): em - ellipse(at right 80em bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): em - ellipse(at right 80em bottom 70em)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): ex - ellipse(at left 50% bottom 70ex)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at top 50% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 50%)" but got "ellipse(at right 1024px top 50%)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70% right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 30%)" but got "ellipse(at right 1024px top 30%)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 896px))" but got "ellipse(at left 50% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at bottom 70ex right 80ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) calc(100% - 896px))" but got "ellipse(at right 1024px bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at right 80% bottom 70ex)
  assert_equals: expected "ellipse(at 20% calc(100% - 896px))" but got "ellipse(at left 20% bottom 896px)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 50%)" but got "ellipse(at right 1024px top 50%)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1024px) 30%)" but got "ellipse(at right 1024px top 30%)"
[FAIL] test unit (computed): ex - ellipse(at right 80ex bottom 70ex)
  assert_equals: expected "ellipse(at calc(100% - 1024px) calc(100% - 896px))" but got "ellipse(at right 1024px bottom 896px)"
[FAIL] test unit (computed): ch - ellipse(at left 50% bottom 70ch)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at top 50% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70% right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at bottom 70ch right 80ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at right 80% bottom 70ch)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): ch - ellipse(at right 80ch bottom 70ch)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at left 50% bottom 70rem)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at top 50% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70% right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 1120px))" but got "ellipse(at left 50% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at bottom 70rem right 80rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at right 80% bottom 70rem)
  assert_equals: expected "ellipse(at 20% calc(100% - 1120px))" but got "ellipse(at left 20% bottom 1120px)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem top 50%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 50%)" but got "ellipse(at right 1280px top 50%)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 1280px) 30%)" but got "ellipse(at right 1280px top 30%)"
[FAIL] test unit (computed): rem - ellipse(at right 80rem bottom 70rem)
  assert_equals: expected "ellipse(at calc(100% - 1280px) calc(100% - 1120px))" but got "ellipse(at right 1280px bottom 1120px)"
[FAIL] test unit (computed): vw - ellipse(at left 50% bottom 70vw)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at top 50% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70% right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at bottom 70vw right 80vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at right 80% bottom 70vw)
  assert_equals: expected "ellipse(at 20% calc(100% - 560px))" but got "ellipse(at left 20% bottom 560px)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw top 50%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vw - ellipse(at right 80vw bottom 70vw)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[FAIL] test unit (computed): vh - ellipse(at left 50% bottom 70vh)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at top 50% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70% right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at bottom 70vh right 80vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at right 80% bottom 70vh)
  assert_equals: expected "ellipse(at 20% calc(100% - 420px))" but got "ellipse(at left 20% bottom 420px)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh top 50%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vh - ellipse(at right 80vh bottom 70vh)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at left 50% bottom 70vmin)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at top 50% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70% right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 420px))" but got "ellipse(at left 50% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at bottom 70vmin right 80vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at right 80% bottom 70vmin)
  assert_equals: expected "ellipse(at 20% calc(100% - 420px))" but got "ellipse(at left 20% bottom 420px)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin top 50%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 50%)" but got "ellipse(at right 480px top 50%)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 480px) 30%)" but got "ellipse(at right 480px top 30%)"
[FAIL] test unit (computed): vmin - ellipse(at right 80vmin bottom 70vmin)
  assert_equals: expected "ellipse(at calc(100% - 480px) calc(100% - 420px))" but got "ellipse(at right 480px bottom 420px)"
[FAIL] test unit (computed): vmax - ellipse(at left 50% bottom 70vmax)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at top 50% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70% right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax left 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax right 50%)
  assert_equals: expected "ellipse(at 50% calc(100% - 560px))" but got "ellipse(at left 50% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at bottom 70vmax right 80vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at right 80% bottom 70vmax)
  assert_equals: expected "ellipse(at 20% calc(100% - 560px))" but got "ellipse(at left 20% bottom 560px)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax top 50%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 50%)" but got "ellipse(at right 640px top 50%)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax bottom 70%)
  assert_equals: expected "ellipse(at calc(100% - 640px) 30%)" but got "ellipse(at right 640px top 30%)"
[FAIL] test unit (computed): vmax - ellipse(at right 80vmax bottom 70vmax)
  assert_equals: expected "ellipse(at calc(100% - 640px) calc(100% - 560px))" but got "ellipse(at right 640px bottom 560px)"
Harness: the test ran to completion.
